Move APIMeta out to it's own new class called ResponseMeta
Note that all refactoring comes in the next commit titled "Document All The Things"
