+++ app/src/main/java/justbe/mindfulnessapp/rest/ResponseWrapper.java
@@ -67,51 +67,5 @@ public class ResponseWrapper <T> {
         return this.error;
     }
 
-    public class APIMeta {
-        private int limit;
-        private String next;
-        private int offset;
-        private String previous;
-        private int total_count;
-
-        public int getLimit() {
-            return limit;
-        }
-
-        public void setLimit(int limit) {
-            this.limit = limit;
-        }
-
-        public String getNext() {
-            return next;
-        }
-
-        public void setNext(String next) {
-            this.next = next;
-        }
-
-        public int getOffset() {
-            return offset;
-        }
-
-        public void setOffset(int offset) {
-            this.offset = offset;
-        }
-
-        public String getPrevious() {
-            return previous;
-        }
-
-        public void setPrevious(String previous) {
-            this.previous = previous;
-        }
-
-        public int getTotal_count() {
-            return total_count;
-        }
-
-        public void setTotal_count(int total_count) {
-            this.total_count = total_count;
-        }
     }
 }
